requirements.txt
transformers>=4.30.0
langchain>=0.0.200
openai>=0.27.0
python-dotenv>=1.0.0
streamlit>=1.24.0
faiss-cpu>=1.7.4
